Add UnixTimestampDateTimeAdapter
Creates a new DateTimeAdapter implementation which uses unix timestamps
(the number of milliseconds that have elapsed since January 1, 1970)
in the form of `number` as NgModel value type.

Also moves some constants and utility methods in native-date-time-adapter
to a common place since this adapter's implementation is very similar to
native-date-time-adapter.

Further performance improvements may be done in the future.
Some of the methods contain multiple date-number conversions.